delete now works, add js to delete or show meal

+++ pages/homepage/goals.js
@@ -861,15 +861,19 @@ Page({
     setDailyCharts(page)
   },
   goToRecipe(e){
-    clearAllChart()
-    wx.pageScrollTo({
-      scrollTop: 0,
-      duration: 0
-    })
-    const id = e.currentTarget.dataset.id
-    wx.navigateTo({
-      url: `/pages/recipes/recipes?id=${id}&showdetail=true&showreview=false`,
-    })
+    const deletebutton = e.target.id == "delete"
+    console.log(deletebutton)
+    if(!deletebutton){
+        clearAllChart()
+        wx.pageScrollTo({
+          scrollTop: 0,
+          duration: 0
+        })
+      const id = e.currentTarget.dataset.id
+      wx.navigateTo({
+        url: `/pages/recipes/recipes?id=${id}&showdetail=true&showreview=false`,
+      })
+    }
   },
   deleteMeal(e){
     clearAllChart()
@@ -879,22 +883,31 @@ Page({
     })
     console.log(e)
     const id = e.currentTarget.dataset.id
-
-    wx.request({
-      url:  `${app.globalData.baseUrl}/meals/${id}`, 
-      header: app.globalData.header,
-      method: 'DELETE',
-      data: {
-        meal: id
-      },
-      success: (res) => {
-        wx.showToast({
-          title: 'The meal is deleted',
-          duration: 1500
-        })
-      },
-      fail: (err) => {
-        console.log(err);
+    wx.showModal({
+      title: 'Delete Meal',
+      content: 'Do you want to continue?',
+      complete: (res) => {
+        if (res.confirm) {
+          wx.request({
+            url:  `${app.globalData.baseUrl}/meals/${id}`, 
+            header: app.globalData.header,
+            method: 'DELETE',
+            data: {
+              meal: id
+            },
+            success: (res) => {
+              console.log(res)
+            },
+            complete(){
+              wx.navigateTo({
+                url: '/pages/homepage/goals?showdaily=true&showweekly=false',
+              })
+            },
+            fail: (err) => {
+              console.log(err);
+            }
+          })
+        }
       }
     })
   },
